refactor(test): DRY and streamline e2e tests
