+++ tests/e2e/options.spec.ts
@@ -1,129 +1,13 @@
-import { BrowserContext, Page } from '@playwright/test'
-import { expect, test } from './fixtures'
+import { expect, test, TestHelpers } from './fixtures'
+
+/**
+ * Tests specific to the Options page functionality that aren't covered by
+ * other consolidated test files (analytics-consent.spec.ts, debug-mode.spec.ts).
+ */
 
 test.describe('Options page', () => {
-  const openOptionsPage = async (context: BrowserContext, extensionId: string): Promise<Page> => {
-    const page = await context.newPage()
-    await page.goto(`chrome-extension://${extensionId}/options.html`)
-    return page
-  }
-
-  test('prompts for analytics consent and persists accept decision', async ({
-    context,
-    extensionId,
-    serviceWorker,
-  }) => {
-    const page = await openOptionsPage(context, extensionId)
-
-    // The consent modal should appear on first visit.
-    const acceptButton = page.getByRole('button', { name: /accept/i })
-    await expect(acceptButton).toBeVisible()
-
-    await acceptButton.click()
-    await expect(acceptButton).toBeHidden()
-
-    // Verify chrome.storage.sync.analytics_consent is true
-    const consent = await serviceWorker.evaluate(async () => {
-      const { analytics_consent } = await chrome.storage.sync.get('analytics_consent')
-      return analytics_consent
-    })
-
-    expect(consent).toBe(true)
-  })
-
-  test('prompts for analytics consent and persists decline decision', async ({
-    context,
-    extensionId,
-    serviceWorker,
-  }) => {
-    const page = await openOptionsPage(context, extensionId)
-
-    // The consent modal should appear on first visit.
-    const declineButton = page.getByRole('button', { name: /decline/i })
-    await expect(declineButton).toBeVisible()
-
-    await declineButton.click()
-    await expect(declineButton).toBeHidden()
-
-    // Verify chrome.storage.sync.analytics_consent is false
-    const consent = await serviceWorker.evaluate(async () => {
-      const { analytics_consent } = await chrome.storage.sync.get('analytics_consent')
-      return analytics_consent
-    })
-
-    expect(consent).toBe(false)
-  })
-
-  test('allows unlocking and toggling debug mode which persists across reloads', async ({
-    context,
-    extensionId,
-    serviceWorker,
-  }) => {
-    const page = await openOptionsPage(context, extensionId)
-
-    // Dismiss consent modal
-    await serviceWorker.evaluate(() => {
-      chrome.storage.sync.set({ analytics_consent: false })
-    })
-
-    // Unlock hidden debug switch by clicking the heading 5×.
-    const heading = page.getByRole('heading', { name: /settings/i })
-    for (let i = 0; i < 5; i++) {
-      await heading.click()
-    }
-
-    const debugSwitch = page.getByRole('switch', { name: /debug mode/i })
-    await expect(debugSwitch).toBeVisible()
-    await expect(debugSwitch).toHaveAttribute('data-state', 'unchecked')
-
-    await debugSwitch.click()
-    await expect(debugSwitch).toHaveAttribute('data-state', 'checked')
-
-    // Verify storage mutation (chrome.storage.local.debugMode === true).
-    const debugMode = await serviceWorker.evaluate(async () => {
-      const { debugMode } = await chrome.storage.local.get('debugMode')
-      return debugMode
-    })
-    expect(debugMode).toBe(true)
-
-    // Reload the page — the switch should remain enabled.
-    await page.reload()
-    const debugSwitchAfterReload = page.getByRole('switch', { name: /debug mode/i })
-    await expect(debugSwitchAfterReload).toHaveAttribute('data-state', 'checked')
-  })
-
-  test('debug row hidden by default and appears after 5 header clicks', async ({
-    context,
-    extensionId,
-    serviceWorker,
-  }) => {
-    // Reset debug mode and unlock debug row in storage to simulate a fresh state
-    await serviceWorker.evaluate(async () => {
-      await chrome.storage.local.set({ debugMode: false, debugUnlocked: false })
-    })
-
-    const page = await context.newPage()
-    await page.goto(`chrome-extension://${extensionId}/options.html`)
-
-    // Dismiss consent modal if it shows up
-    const declineButton = page.getByRole('button', { name: /decline/i })
-    if (await declineButton.isVisible().catch(() => false)) {
-      await declineButton.click()
-      await expect(declineButton).toBeHidden()
-    }
-
-    // Debug row should NOT be present initially
-    const debugSwitch = page.getByRole('switch', { name: /debug mode/i })
-    await expect(debugSwitch).toHaveCount(0)
-
-    // Click the Settings header 5 times to unlock
-    const heading = page.getByRole('heading', { name: /settings/i })
-    for (let i = 0; i < 5; i++) {
-      await heading.click()
-    }
-
-    // Now the debug row should be visible
-    await expect(debugSwitch).toHaveCount(1)
-    await expect(debugSwitch).toBeVisible()
+  test('loads correctly and shows expected title', async ({ context, extensionId }) => {
+    const page = await TestHelpers.openOptionsPage(context, extensionId)
+    await expect(page).toHaveTitle('Scrape Similar - Settings')
   })
 })
