+++ tests/e2e/sidepanel.spec.ts
@@ -1,7 +1,11 @@
 import fs from 'fs/promises'
-import { expect, test } from './fixtures'
+import { expect, test, TestHelpers } from './fixtures'
 
-test.describe('Sidepanel', () => {
+/**
+ * Core sidepanel functionality tests that aren't covered by other test files
+ */
+
+test.describe('Sidepanel Core Functionality', () => {
   test('sidepanel shows unsupported URL splash when opened on a non-supported URL', async ({
     context,
     extensionId,
@@ -16,9 +20,7 @@ test.describe('Sidepanel', () => {
     await testPage.goto(`chrome-extension://${extensionId}/options.html`)
 
     // Dismiss consent modal
-    await serviceWorker.evaluate(() => {
-      chrome.storage.sync.set({ analytics_consent: false })
-    })
+    await TestHelpers.dismissAnalyticsConsent(serviceWorker)
 
     await expect(sidePanel.getByText(/unsupported url/i)).toBeVisible()
   })
@@ -36,78 +38,28 @@ test.describe('Sidepanel', () => {
     await testPage.goto('https://one.one.one.one/')
 
     // Dismiss consent modal
-    await serviceWorker.evaluate(() => {
-      chrome.storage.sync.set({ analytics_consent: false })
-    })
+    await TestHelpers.dismissAnalyticsConsent(serviceWorker)
 
     await expect(sidePanel.getByText(/unsupported url/i)).toBeHidden()
   })
 
-  test('prompts for analytics consent and persists accept decision', async ({
-    serviceWorker,
+  test('can configure and scrape data from a webpage', async ({
     openSidePanel,
+    serviceWorker,
+    context,
   }) => {
-    // Open the side-panel on a blank page.
     const sidePanel = await openSidePanel()
 
-    // The consent modal should appear on first visit.
-    const acceptButton = sidePanel.getByRole('button', { name: /accept/i })
-    await expect(acceptButton).toBeVisible()
+    // Prepare data using shared helper
+    await TestHelpers.prepareSidepanelWithData(sidePanel, serviceWorker, context)
 
-    // Accept analytics collection.
-    await acceptButton.click()
-    await expect(acceptButton).toBeHidden()
+    // Verify that data table is visible
+    await expect(sidePanel.getByRole('heading', { name: /extracted data/i })).toBeVisible()
 
-    // Verify chrome.storage.sync.analytics_consent is true.
-    const consent = await serviceWorker.evaluate(async () => {
-      const { analytics_consent } = await chrome.storage.sync.get('analytics_consent')
-      return analytics_consent
-    })
-
-    expect(consent).toBe(true)
-  })
-
-  test('allows unlocking and toggling debug mode which persists across reloads', async ({
-    serviceWorker,
-    openSidePanel,
-  }) => {
-    // Launch the side-panel.
-    const sidePanel = await openSidePanel()
-
-    // Dismiss consent modal
-    await serviceWorker.evaluate(() => {
-      chrome.storage.sync.set({ analytics_consent: false })
-    })
-
-    // Open the settings drawer via the toolbar button.
-    await sidePanel.getByRole('button', { name: /settings/i }).click()
-
-    // Unlock hidden debug switch by clicking the drawer heading 5×.
-    const heading = sidePanel.getByRole('heading', { name: /settings/i })
-    for (let i = 0; i < 5; i++) {
-      await heading.click()
-    }
-
-    const debugSwitch = sidePanel.getByRole('switch', { name: /debug mode/i })
-    await expect(debugSwitch).toBeVisible()
-    await expect(debugSwitch).toHaveAttribute('data-state', 'unchecked')
-
-    // Enable debug mode.
-    await debugSwitch.click()
-    await expect(debugSwitch).toHaveAttribute('data-state', 'checked')
-
-    // Confirm storage mutation (chrome.storage.local.debugMode === true).
-    const debugMode = await serviceWorker.evaluate(async () => {
-      const { debugMode } = await chrome.storage.local.get('debugMode')
-      return debugMode
-    })
-    expect(debugMode).toBe(true)
-
-    // Reload the side-panel and reopen the settings drawer — the switch should remain enabled.
-    await sidePanel.reload()
-    await sidePanel.getByRole('button', { name: /settings/i }).click()
-    const debugSwitchAfterReload = sidePanel.getByRole('switch', { name: /debug mode/i })
-    await expect(debugSwitchAfterReload).toHaveAttribute('data-state', 'checked')
+    // Verify that we have some data rows
+    const dataRows = sidePanel.locator('tbody tr')
+    const rowCount = await dataRows.count()
+    expect(rowCount).toBeGreaterThan(0)
   })
 
   test('shows error badge for invalid XPath selector', async ({
@@ -121,16 +73,14 @@ test.describe('Sidepanel', () => {
     await testPage.goto('https://en.wikipedia.org/wiki/Playwright_(software)')
 
     // Dismiss consent modal
-    await serviceWorker.evaluate(() => {
-      chrome.storage.sync.set({ analytics_consent: false })
-    })
+    await TestHelpers.dismissAnalyticsConsent(serviceWorker)
 
-    // Type an obviously invalid XPath selector and commit with Enter.
+    // Type an obviously invalid XPath selector and commit with Enter
     const mainSelector = sidePanel.locator('#mainSelector')
     await mainSelector.fill('//*[')
     await mainSelector.press('Enter')
 
-    // Expect a destructive badge with an alert icon (svg) to appear.
+    // Expect a destructive badge with an alert icon (svg) to appear
     const errorBadge = sidePanel.locator('[data-slot="badge"] svg')
     await expect(errorBadge).toBeVisible({ timeout: 1000 })
   })
@@ -146,16 +96,14 @@ test.describe('Sidepanel', () => {
     await testPage.goto('https://en.wikipedia.org/wiki/Playwright_(software)')
 
     // Dismiss consent modal
-    await serviceWorker.evaluate(() => {
-      chrome.storage.sync.set({ analytics_consent: false })
-    })
+    await TestHelpers.dismissAnalyticsConsent(serviceWorker)
 
-    // Provide a valid selector that should match many elements.
+    // Provide a valid selector that should match many elements
     const mainSelector = sidePanel.locator('#mainSelector')
     await mainSelector.fill('//p')
     await mainSelector.press('Enter')
 
-    // Expect a badge whose text is a positive integer to appear.
+    // Expect a badge whose text is a positive integer to appear
     const countBadge = sidePanel.locator('[data-slot="badge"]').filter({ hasText: /^\d+$/ })
     await expect(countBadge).toBeVisible({ timeout: 5000 })
     const badgeText = await countBadge.textContent()
@@ -174,9 +122,7 @@ test.describe('Sidepanel', () => {
     await testPage.goto('https://en.wikipedia.org/wiki/Playwright_(software)')
 
     // Dismiss consent modal
-    await serviceWorker.evaluate(() => {
-      chrome.storage.sync.set({ analytics_consent: false })
-    })
+    await TestHelpers.dismissAnalyticsConsent(serviceWorker)
 
     // Enter selector that matches multiple elements and commit.
     const input = sidePanel.locator('#mainSelector')
@@ -185,16 +131,14 @@ test.describe('Sidepanel', () => {
 
     // Wait for match-count badge to appear confirming highlight is done.
     const countBadge = sidePanel.locator('[data-slot="badge"]').filter({ hasText: /^\d+$/ })
-    await expect(countBadge).toBeVisible({ timeout: 5000 })
+    await expect(countBadge).toBeVisible()
 
     // Click the Scrape button.
     const scrapeBtn = sidePanel.getByRole('button', { name: /^scrape$/i })
     await scrapeBtn.click()
 
     // Data table heading should appear.
-    await expect(sidePanel.getByRole('heading', { name: /extracted data/i })).toBeVisible({
-      timeout: 50000,
-    })
+    await expect(sidePanel.getByRole('heading', { name: /extracted data/i })).toBeVisible()
   })
 
   test('shows "0 found" state when scrape yields no results', async ({
@@ -204,32 +148,30 @@ test.describe('Sidepanel', () => {
   }) => {
     const sidePanel = await openSidePanel()
 
-    // Open a real page.
+    // Open a real page
     const testPage = await context.newPage()
     await testPage.goto('https://en.wikipedia.org/wiki/Playwright_(software)')
 
-    // Bring side-panel to front.
+    // Bring side-panel to front
     await sidePanel.bringToFront()
 
     // Dismiss consent modal
-    await serviceWorker.evaluate(() => {
-      chrome.storage.sync.set({ analytics_consent: false })
-    })
+    await TestHelpers.dismissAnalyticsConsent(serviceWorker)
 
-    // Enter selector that matches nothing.
+    // Enter selector that matches nothing
     const input = sidePanel.locator('#mainSelector')
     await input.fill('//*[@id="nonexistent_element_for_test"]')
     await input.press('Enter')
 
-    // Wait for highlight badge (will show 0).
+    // Wait for highlight badge (will show 0)
     const zeroBadge = sidePanel.locator('[data-slot="badge"]').filter({ hasText: /^0$/ })
     await expect(zeroBadge).toBeVisible({ timeout: 5000 })
 
-    // Click Scrape.
+    // Click Scrape
     const scrapeBtn = sidePanel.getByRole('button', { name: /^scrape$/i })
     await scrapeBtn.click()
 
-    // Expect button text to show "0 found".
+    // Expect button text to show "0 found"
     await expect(sidePanel.getByRole('button', { name: /0 found/i })).toBeVisible({ timeout: 1500 })
   })
 
@@ -240,26 +182,24 @@ test.describe('Sidepanel', () => {
   }) => {
     const sidePanel = await openSidePanel()
 
-    // Open real page for highlight/scrape logic.
+    // Open real page for highlight/scrape logic
     const testPage = await context.newPage()
     await testPage.goto('https://en.wikipedia.org/wiki/Playwright_(software)')
 
     // Dismiss consent modal
-    await serviceWorker.evaluate(() => {
-      chrome.storage.sync.set({ analytics_consent: false })
-    })
+    await TestHelpers.dismissAnalyticsConsent(serviceWorker)
 
-    // Configure selector.
+    // Configure selector
     const selectorValue = '//h2'
     const mainSelector = sidePanel.locator('#mainSelector')
     await mainSelector.fill(selectorValue)
     await mainSelector.press('Enter')
 
-    // Wait for valid badge so Save button becomes enabled.
+    // Wait for valid badge so Save button becomes enabled
     const countBadge = sidePanel.locator('[data-slot="badge"]').filter({ hasText: /^\d+$/ })
     await expect(countBadge).toBeVisible({ timeout: 5000 })
 
-    // Open Save Preset drawer.
+    // Open Save Preset drawer
     await sidePanel
       .getByRole('button', { name: /^save$/i })
       .first()
@@ -269,23 +209,23 @@ test.describe('Sidepanel', () => {
     const presetNameInput = sidePanel.getByPlaceholder('Preset name')
     await presetNameInput.fill(presetName)
 
-    // Click inner Save button.
+    // Click inner Save button
     await sidePanel.getByRole('button', { name: /^save$/i }).click()
 
-    // Wait for drawer to close (input hidden).
+    // Wait for drawer to close (input hidden)
     await expect(presetNameInput).toBeHidden({ timeout: 5000 })
 
-    // Clear the main selector input.
+    // Clear the main selector input
     await mainSelector.clear()
 
-    // Open Load combobox.
+    // Open Load combobox
     await sidePanel.getByRole('button', { name: /load/i }).click()
 
-    // Select the preset.
+    // Select the preset
     const loadItem = sidePanel.getByRole('option', { name: new RegExp(presetName, 'i') })
     await loadItem.click()
 
-    // After load the main selector input should contain the preset's selector.
+    // After load the main selector input should contain the preset's selector
     await expect(mainSelector).toHaveValue(selectorValue)
   })
 
@@ -296,39 +236,13 @@ test.describe('Sidepanel', () => {
   }) => {
     const sidePanel = await openSidePanel()
 
-    // Navigate to real page for scraping.
-    const livePage = await context.newPage()
-    await livePage.goto('https://en.wikipedia.org/wiki/Playwright_(software)')
+    // Prepare data using shared helper
+    await TestHelpers.prepareSidepanelWithData(sidePanel, serviceWorker, context)
 
-    // Dismiss consent modal
-    await serviceWorker.evaluate(() => {
-      chrome.storage.sync.set({ analytics_consent: false })
-    })
-
-    // Prepare selector.
-    const mainSelector = sidePanel.locator('#mainSelector')
-    await mainSelector.fill('//h2')
-    await mainSelector.press('Enter')
-
-    // Press automagic config button
-    await sidePanel
-      .getByRole('button', { name: /auto-generate configuration from selector/i })
-      .click()
-
-    // Wait for badge.
-    const countBadge = sidePanel.locator('[data-slot="badge"]').filter({ hasText: /^\d+$/ })
-    await expect(countBadge).toBeVisible({ timeout: 5000 })
-
-    // Scrape.
-    await sidePanel.getByRole('button', { name: /^scrape$/i }).click()
-    await expect(sidePanel.getByRole('heading', { name: /extracted data/i })).toBeVisible({
-      timeout: 5000,
-    })
-
-    // Open Export dropdown.
+    // Open Export dropdown
     await sidePanel.getByRole('button', { name: /export/i }).click()
 
-    // Initiate CSV download and wait for it.
+    // Initiate CSV download and wait for it
     const [download] = await Promise.all([
       sidePanel.waitForEvent('download'),
       sidePanel.getByRole('menuitem', { name: /save all as csv/i }).click(),
@@ -337,23 +251,25 @@ test.describe('Sidepanel', () => {
     const fileName = download.suggestedFilename()
     expect(fileName.toLowerCase()).toMatch(/\.csv$/)
 
-    // Check the file headers correspond to the table headers.
+    // Check the file headers correspond to the table headers
     const fileData = await download.path()
-    const fileContent = await fs.readFile(fileData, 'utf-8')
-    const headers = fileContent
-      .split('\n')[0]
-      .split(',')
-      .map((header) => header.replace(/"/g, ''))
+    if (fileData) {
+      const fileContent = await fs.readFile(fileData, 'utf-8')
+      const headers = fileContent
+        .split('\n')[0]
+        .split(',')
+        .map((header) => header.replace(/"/g, ''))
 
-    const uiHeaders = await sidePanel.evaluate(() => {
-      const ths = Array.from(document.querySelectorAll('table thead th'))
-      // The first two headers are “#” and “Actions”, which arenʼt included in the CSV.
-      return ths
-        .slice(2)
-        .map((th) => th.textContent?.trim() || '')
-        .filter(Boolean)
-    })
-    expect(headers).toEqual(uiHeaders)
+      const uiHeaders = await sidePanel.evaluate(() => {
+        const ths = Array.from(document.querySelectorAll('table thead th'))
+        // The first two headers are "#" and "Actions", which aren't included in the CSV
+        return ths
+          .slice(2)
+          .map((th) => th.textContent?.trim() || '')
+          .filter(Boolean)
+      })
+      expect(headers).toEqual(uiHeaders)
+    }
   })
 
   test('shows rescrape indicator when config changes after successful scrape', async ({
@@ -368,9 +284,7 @@ test.describe('Sidepanel', () => {
     await testPage.goto('https://en.wikipedia.org/wiki/Playwright_(software)')
 
     // Dismiss consent modal
-    await serviceWorker.evaluate(() => {
-      chrome.storage.sync.set({ analytics_consent: false })
-    })
+    await TestHelpers.dismissAnalyticsConsent(serviceWorker)
 
     // Load first system preset: "Internal (relative) Links"
     await sidePanel.getByRole('button', { name: /load/i }).click()
@@ -445,5 +359,66 @@ test.describe('Sidepanel', () => {
       timeout: 10000,
     })
     await expect(rescrapeIcon).toBeHidden({ timeout: 2000 })
+  })
+
+  test('settings drawer opens and closes correctly', async ({ openSidePanel, serviceWorker }) => {
+    const sidePanel = await openSidePanel()
+
+    // Dismiss consent modal
+    await TestHelpers.dismissAnalyticsConsent(serviceWorker)
+
+    // Settings drawer should be closed initially
+    const settingsDrawer = sidePanel.getByRole('dialog', { name: 'Settings' })
+    await expect(settingsDrawer).toBeHidden()
+
+    // Open settings drawer
+    await sidePanel.getByRole('button', { name: /settings/i }).click()
+    await expect(settingsDrawer).toBeVisible()
+
+    // Close settings drawer by clicking outside
+    await sidePanel.mouse.click(10, 10)
+    await expect(settingsDrawer).toBeHidden()
+  })
+
+  test('copies a single row via row action button', async ({
+    openSidePanel,
+    serviceWorker,
+    context,
+  }) => {
+    const sidePanel = await openSidePanel()
+
+    await TestHelpers.prepareSidepanelWithData(sidePanel, serviceWorker, context)
+    await TestHelpers.stubClipboard(sidePanel)
+
+    // Click the first "Copy this row" button (ensure it is enabled)
+    const copyBtn = sidePanel.getByRole('button', { name: /copy this row/i }).first()
+    await expect(copyBtn).toBeEnabled()
+    await copyBtn.click()
+
+    // Verify clipboard capture
+    const text = await TestHelpers.getCopiedText(sidePanel)
+    expect(text).not.toBeNull()
+    expect(text).toContain('\t') // TSV should contain at least one tab
+    expect(text).not.toContain('\n') // TSV should not contain newlines
+  })
+
+  test('copies entire table via Export → Copy to clipboard', async ({
+    openSidePanel,
+    serviceWorker,
+    context,
+  }) => {
+    const sidePanel = await openSidePanel()
+
+    await TestHelpers.prepareSidepanelWithData(sidePanel, serviceWorker, context)
+    await TestHelpers.stubClipboard(sidePanel)
+
+    // Open Export dropdown and click "Copy to clipboard" option
+    await sidePanel.getByRole('button', { name: /export/i }).click()
+    await sidePanel.getByRole('menuitem', { name: /copy all to clipboard/i }).click()
+
+    // Verify clipboard capture contains multiple lines
+    const text = await TestHelpers.getCopiedText(sidePanel)
+    expect(text).not.toBeNull()
+    expect(text!.split('\n').length).toBeGreaterThan(1) // header + at least one data line
   })
 })
